feat: Stable memory helper functions (#4425)
# Motivation
We need a couple of small helper functions.

# Changes
- When state and the accounts_db are stored in separate memories, it is
necessary to recombine them post-upgrade. As the accounts_db is private,
we need a method to enable this.
- We need to be able to determine the schema of the state overall; there
are two paths depending on whether data is partitioned or not.

# Tests
None; these are both trivial and will be exercised in tests of the
functions that use them.

# Todos

- [ ] Add entry to changelog (if necessary).
  Too minor