NNS1-2847: Use IcrcWalletPage in SnsWallet (#4430)
# Motivation

SnsWallet and IcrcWallet have entirely separate implementations even
though they should have identical functionality.
The reason for this is that SNS universes are identified by the
rootCanisterId while ICRC tokens are identified by the ledgerCanisterId.
So at each level of they component hierarchy, tokens, accounts,
balances, etc. are identified in a different way.
However if we convert the rootCanisterId to the SNS ledger canister ID,
we should be able to reuse the IcrcWalletPage component and the entire
hierarchy below it and then delete a lot of code.

Note: the new `SnsWallet.svelte` is based on `IcrcWallet.svelte` and not
based on the old `SnsWallet.svelte`. So in the first commit, I made
`SnsWallet.svelte` an exact copy of `IcrcWallet.svelte` so that in the
next commit it's easy to see what the changes are relative to
`IcrcWallet.svelte`.

Deleting code that became unused as a result will be done in separate
PRs.

# Changes

1. Make `SnsWallet.svelte` a copy of `IcrcWallet.svelte` but with the
ledger canister ID derived from the root canister ID instead of from the
universe ID.
2. Include `<IcrcTokenAccountsModals />` in `Wallet.svelte` also for SNS
wallets. Previously `SnsWallet` had `SnsTransactionModal` embedded, but
`IcrcWalletPage` requires `IcrcTokenAccountsModals` to handle the
transaction modal.
3. In `sns.services.ts`, put SNS tokens in the `tokensStore` keyed by
ledger canister ID in addition to root canister ID. The tokens keyed by
root canister will be removed in the future PR to make sure this is
properly tested and anything depending on it depends on the ledger
canister ID instead.

# Tests

1. Add `IcrcTokenAccountsModals` in `AccountsTest.svelte` to facilitate
tests that open a transaction modal. Currently it only has
`AccountsModals` which has the receive modals and the ICP transaction
modal, but not the ICRC transaction modal. This wasn't needed before
because SnsWallet had its own transaction modal embedded.
2. Adapt `SnsWallet.spec.ts` to how `IcrcWalletPage` works. Use ICRC
mocks instead of SNS mocks and expecting slightly different behavior
because the implementations of `SnsWallet` and `IcrcWallet` had already
diverged.
3. Add a unit test for `Wallet.spec.ts` to make sure the transaction
modal can be opened for the SNS wallet.
4. Add a unit test to test that SNS tokens are now keyed by ledger
canister ID as well as root canister ID.
5. Update page objects to the new structure.
6. Support setting `ledgerCanisterId` with `setSnsProjects` because the
ledger canister ID is now a lot more important to the implementation.
7. Tested manually ad
https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/

# Todos

- [x] Add entry to changelog (if necessary).