NNS1-2851: Refactor uncertifiedLoadSnsesAccountsBalances (#4444)
# Motivation

Unify wallet implementations.

In this PR, change the implementation of
`uncertifiedLoadSnsAccountsBalances` to use current service
`loadSnsAccounts`.

This way merging api functions and stores later will be easier.

# Changes

* Rename `uncertifiedLoadSnsAccountsBalances` to
`uncertifiedLoadSnsesAccountsBalances`. Notice the "Snses".
* `uncertifiedLoadSnsesAccountsBalances` uses `loadSnsAccounts` and
doesn't load the SNS token because that's done by the aggregator
service.
* Add optional `strategy` param to `loadSnsAccounts`.
* Rename usages of `uncertifiedLoadSnsAccountsBalances`.

# Tests

* Add test for `strategy` param in `laodSnsAccounts` and change expects
on the test that calls `loadSnsAccounts` without `strategy`.

# Todos

- [ ] Add entry to changelog (if necessary).

Not necessary.