NNS1-2849: Remove unused code from SnsIncreaseStakeNeuronModal (#4464)
# Motivation

SnsIncreaseStakeNeuronModal has code to call `loadSnsAccounts` but only
if `transactionFee` or `governanceCanisterId` is not available. Both of
these come from the aggregator data, which is always loaded on app init.
And even if they weren't, calling `loadSnsAccounts` doesn't load the
token or the project.

Previously, it would call `syncSnsAccount`, which would at least load
the SNS token as well, but that was changed because we always have the
SNS token from the aggregator data.

And the accounts are already loaded on the `SnsNeuronDetail` before you
can open the `SnsIncreaseStakeNeuronModal`, and even if they weren't,
the current code would never call `loadSnsAccounts` because it doesn't
check if the accounts are loaded or not, and only checks
`transactionFee` and `governanceCanisterId` which are always there
because they come from the aggregator data.

# Changes

Remove the code that never triggers and that wouldn't be useful even if
it did trigger.

# Tests

Remove the corresponding tests.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary