Remove cron for app deploy (#4473)
# Motivation

We have a workflow to release nns-dapp to an app subnet.
We don't use it often, and when we do, we use the manual dispatch.
Most of the time, the app is out of cycles.
No need to waste cycles on weekly deploys that we don't use.

# Changes

Remove the scheduled deploy and only leave the manual deploy trigger.

# Tests

no

# Todos

- [x] Add entry to changelog (if necessary).