feat: Use State::from instead of restore() (#4525)
# Motivation
In previous PRs, `State::from(memory)` was defined. This recovers state
from memory from either managed virtual memory or raw memory. This
`from(memory)` function supersedes `State::restore()` which recovers
memory only from raw memory.

Note: `restore()` took no arguments as it read the global stable memory
state object. This makes it impossible to test outside a canister, while
the `from(memory)` function can be tested against memory that is passed
in. There is a gap in the logic when using raw memory with the old API,
as that still uses the global object, but the plan is to delete that
code path once we no longer use raw memory.

# Changes
- Replace the call to `restore()` with one to `from(memory)`
- Delete now unused code.

# Tests
- There is an existing e2e test that verifies that this canister can be
downgraded to current prod and back again.
- The `from(memory)` function already exists and has unit tests for
managed memory. For raw memory, we have to rely on the e2e tests.

# Todos

- [x] Add entry to changelog (if necessary).