Fix flaky merge neurons e2e test (#4547)
# Motivation

The `merge-neurons` e2e test became flaky after this recent
https://github.com/dfinity/nns-dapp/pull/4542

https://github.com/dfinity/nns-dapp/actions/runs/8016517254/job/21898719819?pr=4545

What happened is that we click on the "Internet Computer" row in the
tokens table before the balance is loaded.
Even if we explicitly wait for the balance to be loaded, this didn't fix
the issue because `waitForBalance` waits for the spinner to be absent
but it could be called before the spinner has even appeared.

# Changes

1. Fix `TokenTableRow.waitForBalance` to make sure that the element that
would hold the spinner is rendered before waiting for the spinner to be
absent.
2. Override `TokensTableRow.click` to make sure the balance is rendered
before clicking.

# Tests

I was able to reproduce the failure locally and with the fix it passed
20 times in a row.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary