Add e2e test for canisters (#4549)
# Motivation

This is prompted by the change to read transactions from the index
canister so we want to make sure we still render transaction such as
"Create Canister" and "Top-up Canister" correctly.
But we did not yet have any end-to-end test for canisters so I added
more than just the minimum to check transactions.

# Changes

1. Add an end-to-end test which creates a canister, renames a canister
and tops-up a canister.
2. Add missing test IDs.
3. Fix one typo in an existing test ID.
4. Add necessary page objects.
5. Factor out a function to navigate to the NNS main account and reuse
it.
6. Change `typeText` on the Playwright page object to use `fill()`
instead of `type()`. Fill replaces existing text, which we need because
when renaming the canister, the input field is prefilled with the
existing name. Since nothing broke, I assume we never relied on
`typeText()` leaving the existing text there.

# Tests

added

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary