Add icpAccountDetailsStore (#4561)
# Motivation

The current `icpAccountStore` holds a combination of accounts and
balances in a single store. This means all the data has to be loaded
together and if we just want to change a balances, we have to manipulate
the data in the store, which can interfere with loading the accounts.

The cleaner way is to have a separate store for accounts and a separate
store for balances and a derived store which merged both into accounts
with balances. That way when we update a balance in the balances store,
it will just automatically recreate the accounts with balances.

In this PR, I'm only adding a new store for `AccountDetails` as they are
loaded from the nns-dapp. In a later PR this will be use as described
above.

# Changes

Add `icpAccountDetailsStore`.

# Tests

Add unit tests for `icpAccountDetailsStore`.
Tested manually in a separate branch.

# Todos

- [ ] Add entry to changelog (if necessary).
not yet