feat: Notify slack on failed reproduction (#4568)
# Motivation
We would like to report failures in reproducible docker builds in slack.

# Changes
- Rename the title of the docker build workflow.
- At the time it was written there was only one reproducible builds test
but now there are two, leading to confusion.
- Add a job to the workflow that checks that the other jobs complete
successfully. If not, it posts a message to slack.

# Tests
Tested manually thus:
- Injected an error into the Dockerfile
- Forced the workflow to run (by pushing to a branch specified as a push
trigger in the workflow)
- Observed:
- All 4 docker builds completed. The docker build step did not bail out
as soon as one build failed. This is as before, is expected and desired.
If builds fail on one platform but succeed on others, this is useful
information.
- The hash comparison job was skipped. This is unchanged and is as
desired.
- The final check job ran and [posted a message to
slack](https://dfinity.slack.com/archives/C05G2RZFJAJ/p1709131039360009).

# Todos

- [x] Add entry to changelog (if necessary).