NNS1-2857: Make icpAccountBalancesStore a queued store (#4571)
# Motivation

We are trying to replace `icpAccountsStore` with a store that's derived
from `icpAccountDetailsStore` and `icpAccountBalancesStore`.

`icpAccountsStore` is currently a `queuedStore` which allows it to order
queries and updates in the correct order such that old data does not
override new data when a new query is made before an old update
finishes.

I thought with the simplification brought be the refactoring we would no
longer need to use queued stores here but on more careful consideration
it will still be possible (especially in fast e2e tests) that an old
balance result comes back after a new balance has been queried.

So the new `icpAccountBalancesStore` should also be a `queudStore` just
like the old `icpAccountsStore`.

# Changes

Change `icpAccountBalancesStore` to be a `queuedStore`.
See the current queued
[icpAccountsStore](https://github.com/dfinity/nns-dapp/blob/85b176bfaed747e470e72d1852b2c880569a33ef/frontend/src/lib/stores/icp-accounts.store.ts#L47)
for comparison.

# Tests

Unit test have been updated.

# Todos

- [ ] Add entry to changelog (if necessary).
not yet used