NNS1-2857: Make NnsWallet.spec.ts more realistic (#4581)
# Motivation

While testing reloading the balance when opening the wallet, I noticed
that transactions were being loaded twice in certain cases.
One such case is when clicking "Finish" in the receive modal.
Surprisingly we had a test for this that didn't catch this.
It turned out the test finished before the additional transaction
loading happened.
So I decided to improve the test.

We should fix the code to avoid the additional call but this PR only
changes the test to demonstrates the current behavior before we fix it.

# Changes

1. Change `pauseQueryAccountBalance` and `pauseGetTransactions` so it's
allowed to pass a result when resolving the call.
2. Add a missing test that checks that we load (but not too often)
balance and transactions when opening the wallet.
3. Update `"should reload account after finish receiving tokens"` to
check that the new balance and transactions are rendered. This also
reveals the additional call to `getTransactions`.

# Tests

improved

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary